+++ src/core.js
@@ -2,7 +2,7 @@ export default {
     /**
      * Constants
      */
-    VERSION: '0.3.0',
+    VERSION: '0.3.1',
     PREFIX: 'qsl-',
     FLOW_TYPE: {
         DEFAULT: 'default',